+++ gehomesdk/erd/values/common/__init__.py
@@ -1,6 +1,7 @@
 from .erd_appliance_type import ErdApplianceType
 from .erd_brand import ErdBrand, ERD_BRAND_NAME_MAP
 from .erd_clock_format import ErdClockFormat
+from .erd_demand_response_state import ErdDemandResponseState
 from .erd_end_tone import ErdEndTone
 from .erd_locked import ErdInterfaceLocked
 from .erd_measurement_units import ErdMeasurementUnits
@@ -8,3 +9,6 @@ from .erd_on_off import ErdOnOff
 from .erd_personality import ErdPersonality
 from .erd_present import ErdPresent
 from .erd_sound_level import ErdSoundLevel
+from .erd_resource_measurement_type import ErdResourceMeasurementType
+from .erd_gas_type import ErdGasType
+from .erd_dsm_power_usage import ErdDsmPowerUsage